+++ src/main/kotlin/dev/mayaqq/estrogen/compat/recipeviewers/api/CRVPlugin.kt
@@ -8,5 +8,6 @@ interface CRVPlugin {
     val removedRecipes: List<ResourceLocation>
     val recipes: List<ViewerInfo<*, *>>
 
-    val pseudoRecipes: List<PseudoRecipeHolder<*>>
+    val pseudoRecipes: List<PseudoRecipeHolder<*>>?
+        get() = null
 }