+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - Cog Icon in Configs
 - Fixed redash when effect runs out
 - Removed ugly log message on server message
+- Fixed forced pseudo recipes